split off bisecotr files
