+++ tests/unit/file_cache/basic_file_cache_test.py
@@ -7,17 +7,11 @@ from buckaroo.file_cache.base import (
     FileCache,
     Executor,
     ProgressNotification,
-    ExpressionBisector,
-    ColumnBisector,
-    RowRangeBisector,
-    get_columns_from_args,
-    ExecutorLogEvent,
-    SimpleExecutorLog,
 )
 import polars as pl
 import polars.selectors as cs
 from typing import cast
-from datetime import timedelta, datetime as dtdt
+from datetime import timedelta
 # fc = FileCache()    
 
 # def pseudo(fname:str) -> None:
@@ -202,18 +196,6 @@ df = pl.DataFrame({
     'b2': ["foo", "bar", "baz"]
     })
 ldf = df.lazy()
-        
-def _expr_labels(exprs:list[pl.Expr]) -> set[str]:
-    labels:set[str] = set()
-    for e in exprs:
-        s = str(e)
-        if 'sum()' in s:
-            labels.add('sum')
-        elif 'count()' in s:
-            labels.add('len')
-        elif 'hash_series' in s or 'hash_xx' in s:
-            labels.add('hash')
-    return labels
 
 def test_simple_executor():
 
@@ -268,95 +250,7 @@ def test_simple_executor_on_fail():
 
     assert exc.executor_log.check_log_for_previous_failure(exc.dfi, ev.args) == True
 
-def test_bisect():
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-
-    # produce a failure event (on _sum)
-    exc = Executor(ldf, FailOnSumExecutor(), listener, fc)
-    exc.run()
-    evs = exc.executor_log.get_log_events()
-    failing_events = [ev for ev in evs if ev.completed == False]
-    assert len(failing_events) >= 1
-    fail_input = failing_events[0]
-
-    # run bisector to minimize failing expressions and maximize success
-    bi = ExpressionBisector(fail_input, exc.executor_log, FailOnSumExecutor(), ldf)
-    fail_ev, success_ev = bi.run()
-
-    assert fail_ev.completed == False
-    assert len(fail_ev.args.expressions) == 1
-    # verify the failing expr is the sum expression (column name ends with _sum)
-    fail_cols = get_columns_from_args(ldf, fail_ev.args)
-    assert fail_cols == ['a1_sum']
-    assert _expr_labels(fail_ev.args.expressions) == {'sum'}
-
-    assert success_ev.completed == True
-    # verify the success exprs are hash and len
-    succ_cols = set(get_columns_from_args(ldf, success_ev.args))
-    assert succ_cols == {'a1_hash', 'a1_len'}
-    assert _expr_labels(success_ev.args.expressions) == {'hash','len'}
-
-
-def test_bisector_multiple_failing_expressions():
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-
-    exc = Executor(ldf, FailOnHashOrSumExecutor(), listener, fc)
-    exc.run()
-    evs = exc.executor_log.get_log_events()
-    failing_events = [ev for ev in evs if ev.completed == False]
-    assert len(failing_events) >= 1
-    fail_input = failing_events[0]
-
-    bi = ExpressionBisector(fail_input, exc.executor_log, FailOnHashOrSumExecutor(), ldf)
-    fail_ev, success_ev = bi.run()
-
-    # minimal failing set should be one expression (either hash or sum)
-    assert fail_ev.completed == False
-    assert len(fail_ev.args.expressions) == 1
-    fail_cols = set(get_columns_from_args(ldf, fail_ev.args))
-    assert fail_cols in ({'a1_hash'}, {'a1_sum'})
-    assert _expr_labels(fail_ev.args.expressions) in ({'hash'}, {'sum'})
-
-    # maximal success should then be the remaining safe expression(s); given both hash and sum fail,
-    # only len should remain
-    assert success_ev.completed == True
-    succ_cols = set(get_columns_from_args(ldf, success_ev.args))
-    assert succ_cols == {'a1_len'}
-    assert _expr_labels(success_ev.args.expressions) == {'len'}
-
-
-def test_bisector_on_success_event_noop():
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-
-    exc = Executor(ldf, SimpleColumnExecutor(), listener, fc)
-    existing_stats = {'a1':{}, 'b2':{}}
-    starting_args = SimpleColumnExecutor().get_execution_args(existing_stats)  # type: ignore
-    success_input = ExecutorLogEvent(
-        dfi=exc.dfi,
-        args=starting_args,
-        start_time=dtdt.now(),
-        end_time=dtdt.now(),
-        completed=True,
-    )
-
-    bi = ExpressionBisector(success_input, exc.executor_log, SimpleColumnExecutor(), ldf)
-    fail_ev, success_ev = bi.run()
-
-    # when starting from a success event, both returned events should be success
-    assert success_ev.completed == True
-    assert fail_ev.completed == True
-    succ_cols = set(get_columns_from_args(ldf, success_ev.args))
-    fail_cols = set(get_columns_from_args(ldf, fail_ev.args))
-    assert succ_cols == {'a1_hash', 'a1_sum', 'a1_len', 'b2_hash', 'b2_len'}
-    assert fail_cols == {'a1_hash', 'a1_sum', 'a1_len', 'b2_hash', 'b2_len'}
-    assert _expr_labels(success_ev.args.expressions) == {'hash','sum','len'}
-    assert _expr_labels(fail_ev.args.expressions) == {'hash','sum','len'}
+ 
 
 def test_simple_executor_listener_calls():
     fc = FileCache()
@@ -448,128 +342,3 @@ def test_in_memory_cache():
 
   """
 
-class FailOnColumnExecutor(SimpleColumnExecutor):
-    """
-      fails if a specific column is present in args.columns
-      """
-    def __init__(self, bad_col: str) -> None:
-        super().__init__()
-        self.bad_col = bad_col
-    def execute(self, ldf:pl.LazyFrame, execution_args:ExecutorArgs) -> ColumnResults:
-        if self.bad_col in execution_args.columns:
-            1/0
-        # behave like parent when not failing
-        return super().execute(ldf, execution_args)
-
-
-def test_column_bisector():
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-    exc = Executor(ldf, FailOnColumnExecutor('a1'), listener, fc)
-    # construct a starting failing event with both columns
-    existing_stats = {'a1':{}, 'b2':{}}
-    starting_args = SimpleColumnExecutor().get_execution_args(existing_stats)  # type: ignore
-    starting_ev = ExecutorLogEvent(
-        dfi=exc.dfi,
-        args=starting_args,
-        start_time=dtdt.now(),
-        end_time=None,
-        completed=False,
-    )
-    bi = ColumnBisector(starting_ev, exc.executor_log, FailOnColumnExecutor('a1'), ldf)
-    fail_ev, success_ev = bi.run()
-    assert fail_ev.completed == False
-    assert success_ev.completed == True
-    assert fail_ev.args.columns == ['a1']
-    assert success_ev.args.columns == ['b2']
-
-
-def test_column_bisector_on_success_event_noop():
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-    exc = Executor(ldf, SimpleColumnExecutor(), listener, fc)
-    existing_stats = {'a1':{}, 'b2':{}}
-    starting_args = SimpleColumnExecutor().get_execution_args(existing_stats)  # type: ignore
-    starting_ev = ExecutorLogEvent(
-        dfi=exc.dfi,
-        args=starting_args,
-        start_time=dtdt.now(),
-        end_time=dtdt.now(),
-        completed=True,
-    )
-    bi = ColumnBisector(starting_ev, exc.executor_log, SimpleColumnExecutor(), ldf)
-    fail_ev, success_ev = bi.run()
-    assert fail_ev.completed == True
-    assert success_ev.completed == True
-    assert fail_ev.args.columns == ['a1','b2']
-    assert success_ev.args.columns == ['a1','b2']
-
-
-class RowRangeAwareFailingExecutor(SimpleColumnExecutor):
-    """
-      Fails if the row range includes [3,7) in terms of args.row_start/row_end
-      (row_end exclusive). If row_start/row_end are None, treat as full range (failing).
-      """
-    def execute(self, ldf:pl.LazyFrame, execution_args:ExecutorArgs) -> ColumnResults:
-        rs = execution_args.row_start
-        re = execution_args.row_end
-        if rs is None or re is None:
-            # consider full window, fail if it includes [3,7)
-            1/0
-        if rs <= 3 and re >= 7:
-            1/0
-        return super().execute(ldf, execution_args)
-
-
-def test_row_range_bisector_minimal_and_success():
-    # Build a larger dataframe for row-range tests
-    df2 = pl.DataFrame({
-        'a1': list(range(10)),
-        'b2': [str(i) for i in range(10)],
-    })
-    ldf2 = df2.lazy()
-    fc = FileCache()
-    def listener(progress:ProgressNotification) -> None:
-        pass
-
-    # starting failing event with both columns and no row range
-    existing_stats = {'a1':{}, 'b2':{}}
-    starting_args = SimpleColumnExecutor().get_execution_args(existing_stats)  # type: ignore
-    starting_ev = ExecutorLogEvent(
-        dfi=(id(ldf2), ''),
-        args=starting_args,
-        start_time=dtdt.now(),
-        end_time=None,
-        completed=False,
-    )
-    rr = RowRangeBisector(starting_ev, SimpleExecutorLog(), RowRangeAwareFailingExecutor(), ldf2)  # type: ignore
-    fail_ev, success_ev = rr.run()
-    assert fail_ev.completed == False
-    assert success_ev.completed == True
-    assert fail_ev.args.row_start == 3
-    assert fail_ev.args.row_end == 7
-    # success should be the larger of [0,3) and [7,10)
-    assert (success_ev.args.row_start, success_ev.args.row_end) in [(0,3), (7,10)]
-
-
-def test_row_range_bisector_on_success_event_noop():
-    df2 = pl.DataFrame({
-        'a1': list(range(10)),
-        'b2': [str(i) for i in range(10)],
-    })
-    ldf2 = df2.lazy()
-    existing_stats = {'a1':{}, 'b2':{}}
-    starting_args = SimpleColumnExecutor().get_execution_args(existing_stats)  # type: ignore
-    success_ev = ExecutorLogEvent(
-        dfi=(id(ldf2), ''),
-        args=starting_args,
-        start_time=dtdt.now(),
-        end_time=dtdt.now(),
-        completed=True,
-    )
-    rr = RowRangeBisector(success_ev, SimpleExecutorLog(), RowRangeAwareFailingExecutor(), ldf2)  # type: ignore
-    fev, sev = rr.run()
-    assert fev.completed == True
-    assert sev.completed == True
